Refactored UserData
