+++ src/main/java/gdx/liftoff/Main.java
@@ -29,10 +29,9 @@ import com.kotcrab.vis.ui.widget.file.FileChooser;
 import com.kotcrab.vis.ui.widget.file.FileChooser.SelectionMode;
 import com.kotcrab.vis.ui.widget.file.FileChooserAdapter;
 import com.ray3k.stripe.*;
-import gdx.liftoff.config.Configuration;
 import gdx.liftoff.ui.OverlayTable;
 import gdx.liftoff.ui.RootTable;
-import gdx.liftoff.ui.data.UserData;
+import gdx.liftoff.ui.UserData;
 import gdx.liftoff.ui.dialogs.FullscreenDialog;
 import org.lwjgl.PointerBuffer;
 import org.lwjgl.util.nfd.NativeFileDialog;
