+++ src/main/java/gdx/liftoff/ui/dialogs/PlatformsDialog.java
@@ -10,7 +10,7 @@ import com.ray3k.stripe.CollapsibleGroup;
 import com.ray3k.stripe.CollapsibleGroup.CollapseType;
 import com.ray3k.stripe.PopTable;
 import com.ray3k.stripe.ScaleContainer;
-import gdx.liftoff.ui.data.UserData;
+import gdx.liftoff.ui.UserData;
 
 import static gdx.liftoff.Main.*;
