+++ src/main/java/gdx/liftoff/ui/panels/ProjectPanel.java
@@ -6,9 +6,8 @@ import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.scenes.scene2d.ui.TextField;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.utils.Align;
-import com.github.czyzby.kiwi.util.common.Strings;
 import gdx.liftoff.Main;
-import gdx.liftoff.ui.data.UserData;
+import gdx.liftoff.ui.UserData;
 import gdx.liftoff.ui.dialogs.FullscreenDialog;
 
 import static gdx.liftoff.Main.*;
